item availability update successfully
